+++ ac3dec/libac3/bit_allocate.h
@@ -23,4 +23,4 @@
 
 void bit_allocate(int fscod, audblk_t * audblk, ac3_ba_t * ba, uint16_t start,
 		  uint16_t end, int16_t fastleak, int16_t slowleak,
-		  uint8_t * exp, uint16_t * bap);
+		  uint8_t * exp, int8_t * bap);
